+++ apps/cms-strapi/src/admin/webpack.config.ts
@@ -1,0 +1,8 @@
+import { type Configuration } from "webpack";
+
+export default (config: Configuration) => {
+	return {
+		...config,
+		output: { ...config?.output, hashFunction: "sha256" },
+	};
+};