+++ apps/cms-strapi/src/plugins/uno/admin/src/pages/HomePage/index.tsx
@@ -1,0 +1,20 @@
+/*
+ *
+ * HomePage
+ *
+ */
+
+import React from "react";
+
+import { PLUGIN_ID } from "../../../../utils/constants.js";
+
+const HomePage = () => {
+	return (
+		<div>
+			<h1>{PLUGIN_ID}&apos;s HomePage</h1>
+			<p>Happy coding</p>
+		</div>
+	);
+};
+
+export default HomePage;