+++ apps/cms-strapi/src/plugins/uno/server/controllers/index.ts
@@ -1,0 +1,5 @@
+import { unoController } from "./uno.js";
+
+export default {
+	uno: unoController,
+};