+++ apps/cms-strapi/src/plugins/uno/server/routes/adminRoutes/index.ts
@@ -1,0 +1,15 @@
+import { type AdminRoutes } from "../types.js";
+
+export const adminRoutes = {
+	type: "admin",
+	routes: [
+		{
+			method: "GET",
+			path: "/uno",
+			handler: "uno.index",
+			config: {
+				policies: ["admin::isAuthenticatedAdmin"],
+			},
+		},
+	],
+} satisfies AdminRoutes;