+++ apps/cms-strapi/src/plugins/uno/server/routes/contentApiRoutes/index.ts
@@ -1,0 +1,6 @@
+import { type ContentApiRoutes } from "../types.js";
+
+export const contentApiRoutes = {
+	type: "content-api",
+	routes: [],
+} satisfies ContentApiRoutes;