+++ apps/cms-strapi/src/plugins/uno/server/routes/index.ts
@@ -1,0 +1,8 @@
+import { adminRoutes } from "./adminRoutes/index.js";
+import { contentApiRoutes } from "./contentApiRoutes/index.js";
+import { type Routes } from "./types.js";
+
+export default {
+	admin: adminRoutes,
+	"content-api": contentApiRoutes,
+} satisfies Routes;