+++ apps/strapi-app/src/admin/webpack.config.ts
@@ -1,6 +1,0 @@
-import { type Configuration } from "webpack";
-
-export default (config: Configuration) => ({
-	...config,
-	output: { ...config?.output, hashFunction: "sha256" },
-});