+++ lint-staged.config.js
@@ -3,21 +3,33 @@ import { ESLint } from "eslint";
 export const removeIgnoredFiles = async (files) => {
 	const eslint = new ESLint();
 	const isIgnored = await Promise.all(
-		files.map((file) => eslint.isPathIgnored(file)),
+		files.map((file) => {
+			return eslint.isPathIgnored(file);
+		}),
 	);
-	const filteredFiles = files.filter((_, index) => !isIgnored[index]);
+	const filteredFiles = files.filter((_, index) => {
+		return !isIgnored[index];
+	});
 
 	return filteredFiles.join(" ");
 };
 
 export default {
-	"*": [`yarn lint:prettier`, `yarn lint:spelling`],
-	"**/*.{ts,tsx,d.ts,cts,d.cts,mts,d.mts}": () => "yarn lint:types",
+	"*": "yarn format:prettier",
+	"**/*.md": async (files) => {
+		return [
+			`yarn markdownlint-cli2 --fix ${files} '#**/node_modules/**' '#apps/cms-strapi/.cache/**' '#**/CHANGELOG.md'`,
+		];
+	},
 	"**/*.{json,yml,yaml,toml,js,jsx,cjs,mjs,ts,tsx,all-contributorsrc,code-workspace}":
 		async (files) => {
 			const filesToLint = await removeIgnoredFiles(files);
 
-			return [`yarn format:eslint ${filesToLint}`];
+			return [
+				`yarn eslint --max-warnings 0 --ignore-path .gitignore --cache --report-unused-disable-directives --color --fix ${filesToLint} `,
+			];
 		},
-	"package.json": "yarn lint:package-json",
+	"**/*.{ts,tsx,d.ts,cts,d.cts,mts,d.mts}": () => {
+		return "yarn lint:types";
+	},
 };