+++ packages/strapi-plugin-uno/admin/src/pluginId.ts
@@ -1,9 +1,0 @@
-import pluginPkg from "../../package.json";
-
-const pluginId = pluginPkg.name.replace(
-	// eslint-disable-next-line regexp/no-unused-capturing-group
-	/^(@[^,.-][\w,-]+\/|strapi-)plugin-/iu,
-	"",
-);
-
-export default pluginId;